# docs/swagger.yaml
basePath: /api/v1
definitions:
  Response:
    properties:
      code:
        type: integer
      data: {}
      message: {}
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  model.CreateFile:
    properties:
      cover:
        type: string
      kind:
        type: string
      name:
        type: string
      number:
        type: integer
    type: object
...
  model.DriftingNovel:
    properties:
      cover:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      kind:
        type: integer
      name:
        type: string
      number:
        type: integer
      ownerID:
        type: integer
      theme:
        type: string
      updatedAt:
        type: string
      writer_number:
        type: integer
    type: object
...
  model.Invite:
    properties:
      file_id:
        type: integer
      file_kind:
        type: string
      friend_id:
        type: integer
      host_id:
        type: integer
    required:
    - file_id
    - file_kind
    type: object
  model.JoinedDrifting:
    properties:
      DriftingNovelID:
        type: integer
      driftingDrawingID:
        type: integer
      driftingNoteID:
        type: integer
      driftingPictureID:
        type: integer
      studentID:
        type: integer
    type: object
...
  model.NovelContact:
    properties:
      file_id:
        type: integer
      the_words:
        type: string
      writer_id:
        type: integer
    type: object
  model.State:
    properties:
      file_id:
        type: integer
      file_kind:
        type: string
      writer_id:
        type: integer
    type: object
...
host: 116.204.121.9:61583
info:
  contact:
    email: SJMbaiyang@163.com
    name: KitZhangYs
  description: 漂流API
  termsOfService: http://swagger.io/terms/
  title: Drifting API
  version: "1.0"
paths:
...
  /api/v1/drifting_novel/accept:
    post:
      consumes:
      - application/json
      description: 接受好友创作邀请，注：该接口仅负责删除对应邀请记录，后续操作需调用参与创作接口
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 要通过的邀请
        in: body
        name: TheInvite
        required: true
        schema:
          $ref: '#/definitions/model.Invite'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"Success"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"Failure"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 接受创作邀请
      tags:
      - driftingnovel
  /api/v1/drifting_novel/create:
    get:
      consumes:
      - application/json
      description: 获取对应用户创建的漂流小说
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"获取成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"Failure"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 获取用户漂流小说
      tags:
      - driftingnovel
    post:
      consumes:
      - application/json
      description: 创建漂流小说,kind必备，且只能为"熟人模式"和"生人模式"，否则将无法进行筛选及推送
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 新建漂流小说信息
        in: body
        name: DriftingNovel
        required: true
        schema:
          $ref: '#/definitions/model.CreateFile'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"创建成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"创建失败"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 创建漂流小说
      tags:
      - driftingnovel
  /api/v1/drifting_novel/delete:
    delete:
      consumes:
      - application/json
      description: 删除指定漂流小说
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 要删除的漂流小说
        in: body
        name: ThePicture
        required: true
        schema:
          $ref: '#/definitions/model.DriftingNovel'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"删除成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"删除失败，您有可能不是该文件的主人，或者该文件不存在"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 删除漂流小说
      tags:
      - driftingnovel
  /api/v1/drifting_novel/detail:
    post:
      consumes:
      - application/json
      description: 获取漂流小说内容，需在json中提供漂流小说的ID
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 获取的ID
        in: body
        name: FDriftingNovel
        required: true
        schema:
          $ref: '#/definitions/model.DriftingNovel'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"获取成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"获取失败"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 获取漂流小说内容
      tags:
      - driftingnovel
  /api/v1/drifting_novel/invite:
    get:
      consumes:
      - application/json
      description: 获取该用户的邀请信息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"获取成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"获取信息失败"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 获取邀请信息
      tags:
      - driftingnovel
    post:
      consumes:
      - application/json
      description: 邀请好友创作，需在json中提供好友学号，漂流小说ID，及文件类型(漂流小说需注明是DriftingNovel)
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 新建邀请
        in: body
        name: NewInvite
        required: true
        schema:
          $ref: '#/definitions/model.Invite'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"邀请成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"邀请失败，你可能已邀请过该好友"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 邀请好友进行创作
      tags:
      - driftingnovel
  /api/v1/drifting_novel/join:
    get:
      consumes:
      - application/json
      description: 获取用户参加得漂流小说信息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"获取成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"Failure"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 获取用户参加的漂流小说信息
      tags:
      - driftingnovel
    post:
      consumes:
      - application/json
      description: 参加漂流小说创作
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 要参加的漂流小说
        in: body
        name: Joined
        required: true
        schema:
          $ref: '#/definitions/model.JoinedDrifting'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"参加成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"Failure"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 参加漂流小说创作
      tags:
      - driftingnovel
  /api/v1/drifting_novel/recommendation:
    get:
      consumes:
      - application/json
      description: 随机推荐一个漂流小说
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"获取成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"获取失败"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 随机推荐漂流小说
      tags:
      - driftingnovel
  /api/v1/drifting_novel/refuse:
    post:
      consumes:
      - application/json
      description: 拒绝创作邀请
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 拒绝邀请
        in: body
        name: TheInvite
        required: true
        schema:
          $ref: '#/definitions/model.Invite'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"拒绝成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"拒绝失败"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 拒绝邀请
      tags:
      - driftingnovel
  /api/v1/drifting_novel/write:
    post:
      consumes:
      - application/json
      description: 参与漂流小说创作,需要在json中添加名为
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 写的内容
        in: body
        name: NewContact
        required: true
        schema:
          $ref: '#/definitions/model.NovelContact'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"Success"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"Failure"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 参与漂流小说创作(写内容)
      tags:
      - driftingnovel
...
  /api/v1/lock/get_lock:
    post:
      consumes:
      - application/json
      description: 获取当前上锁人信息，需提供文件ID及文件类型(漂流本/漂流照片/漂流小说/漂流画)
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 获取信息请求
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/model.State'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"获取成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"获取失败"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 或取当前上锁人
      tags:
      - Lock
  /api/v1/lock/lock_off:
    delete:
      consumes:
      - application/json
      description: 解锁漂流文件，需提供文件ID及文件类型(漂流本/漂流照片/漂流小说/漂流画)
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 解锁请求体
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/model.State'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"解锁成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"解锁失败"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 解锁漂流文件
      tags:
      - Lock
  /api/v1/lock/lock_on:
    post:
      consumes:
      - application/json
      description: 需提供文件ID，及文件类型(漂流本/漂流照片/漂流小说/漂流画)，需携带token
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 上锁请求
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/model.State'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"上锁成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"上锁失败"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 对漂流文件上锁
      tags:
      - Lock
...
swagger: "2.0"